Upgraded the Secure_Recv function

+++ src/connection/secure_send.h
@@ -10,6 +10,6 @@ extern int Secure_TestSendSize(size_t len);
 
 extern int Secure_SendSize(Secure_Session *session, size_t len);
 
-extern int Secure_SendMessage(Secure_Session *session, const uint8_t *data, size_t len);
+extern size_t Secure_SendMessage(Secure_Session *session, const uint8_t *data, size_t len);
 
 #endif
